covid19-twitter-search-app/requirements.txt
async-timeout==4.0.3
blinker==1.7.0
bson==0.5.10
cachelib==0.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
dnspython==2.6.1
filelock==3.13.4
Flask==3.0.3
Flask-Caching==2.1.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
packaging==24.0
pipenv==2023.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pymongo==4.6.3
python-dateutil==2.9.0.post0
redis==5.0.3
requests==2.31.0
six==1.16.0
urllib3==2.2.1
virtualenv==20.25.3
Werkzeug==3.0.2
zipp==3.18.1
